requirements.txt
requests
python-dotenv
Jinja2
meraki
